Made a separate class and activity for search.
