+++ app/src/main/AndroidManifest.xml
@@ -13,8 +13,11 @@
         android:theme="@style/Theme.B07Project"
         tools:targetApi="31">
         <activity
+            android:name="com.example.b07Project.SearchActivity"
+            android:exported="true" />
+        <activity
             android:name="com.example.b07Project.RemoveActivity"
-            android:exported="false" />
+            android:exported="true" />
         <activity
             android:name="com.example.b07Project.MainActivity"
             android:exported="true">
