+++ app/src/main/java/com/example/b07Project/RemoveActivity.java
@@ -4,15 +4,8 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
-import android.widget.EditText;
-
-import androidx.activity.EdgeToEdge;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.graphics.Insets;
-import androidx.core.view.ViewCompat;
-import androidx.core.view.WindowInsetsCompat;
-
 import com.example.b07project.R;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
@@ -60,25 +53,8 @@ public class RemoveActivity extends AppCompatActivity {
 
     public void removeDeny(View view) {
 
-        Intent intent = new Intent(RemoveActivity.this, MainActivity.class);
+        Intent intent = new Intent(RemoveActivity.this, SearchActivity.class);
         startActivity(intent);
     }
 
-
-    private void updatePathValue(String path, String subPath){
-
-        mDatabase = FirebaseDatabase.getInstance().getReference(path);
-
-        mDatabase.child(subPath).addListenerForSingleValueEvent(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot snapshot) {
-                Integer value = snapshot.getValue(Integer.class);
-                assert value != null;
-                mDatabase.setValue(value - 1);
-            }
-
-            public void onCancelled(@NonNull DatabaseError error) {
-            }
-        });
-    }
 }
